+++ src/starblaster/Player.java
@@ -3,7 +3,7 @@ package starblaster;
 public class Player extends Entity{
     private int health, maxHealth, score;
     
-    public Player(int x, int y, int vx, int vy, int width, int height, int health, int maxHealth, int score){
+    public Player(double x, double y, double vx, double vy, int width, int height, int health, int maxHealth, int score){
         super(x, y, vx, vy, width, height);
         this.health = health;
         this.maxHealth = maxHealth;
